characters system bug fixed & added abilities if accolade is the commander
