+++ Crusader Wars/armies/battle scale/ArmyProportions.cs
