+++ Crusader Wars/armies/characters/SpecialAbilities.cs
