+++ Crusader Wars/client/RequiredMods/RequiredModsMessage.cs
@@ -18,6 +18,19 @@ namespace Crusader_Wars.client.RequiredMods
             InitializeComponent();
             this.Icon = Properties.Resources.logo;
         }
+        protected override void WndProc(ref Message m)
+        {
+            switch (m.Msg)
+            {
+                case 0x84:
+                    base.WndProc(ref m);
+                    if ((int)m.Result == 0x1)
+                        m.Result = (IntPtr)0x2;
+                    return;
+            }
+
+            base.WndProc(ref m);
+        }
 
         private void SetTitleAndMods(string mapperName, string requiredMods)
         {
