+++ Crusader Wars/data/battle results/BattleResult.cs
